requirements/base.txt
Adafruit-Blinka==8.24.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-dht==4.0.3
adafruit-circuitpython-requests==2.0.2
adafruit-circuitpython-typing==1.9.5
Adafruit-PlatformDetect==3.53.0
Adafruit-PureIO==1.1.11
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.3
djangorestframework==3.14.0
kombu==5.3.2
prompt-toolkit==3.0.39
pyftdi==0.55.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.bme280==0.2.4
RPi.GPIO==0.7.1
six==1.16.0
smbus2==0.4.3
sqlparse==0.4.4
sysv-ipc==1.1.0
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.9
